+++ packages/components/src/box-control/all-input-control.tsx
@@ -91,6 +91,7 @@ export default function AllInputControl( {
 			<FlexedRangeControl
 				__nextHasNoMarginBottom
 				__next40pxDefaultSize={ __next40pxDefaultSize }
+				__shouldNotWarnDeprecated36pxSize
 				aria-controls={ inputId }
 				label={ LABELS.all }
 				hideLabelFromVision