+++ packages/components/src/box-control/input-controls.tsx
@@ -134,6 +134,7 @@ export default function BoxInputControls( {
 						<FlexedRangeControl
 							__nextHasNoMarginBottom
 							__next40pxDefaultSize={ __next40pxDefaultSize }
+							__shouldNotWarnDeprecated36pxSize
 							aria-controls={ inputId }
 							label={ LABELS[ side ] }
 							hideLabelFromVision